functions_with_openai/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pywin32
openai==0.28
openai[datalib]
numpy
azure-cosmos
pymongo